[ELY-339] renaming storage to store
